service_mart/requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.7
django-appconf==1.0.3
django-coverage==1.2.4
djangorestframework==3.9.2
idna==2.8
Markdown==3.0.1
mysqlclient==1.4.2.post1
pip==19.0.3
PyMySQL==0.9.3
pytz==2018.9
requests==2.21.0
setuptools==33.1.1
six==1.12.0
urllib3==1.24.1
uWSGI==2.0.18
wheel==0.33.1
django-redis==4.10.0
ipython-genutils==0.2.0
ipython==7.4.0
